requirements.txt
streamlit>=1.28.0
requests>=2.31.0
plotly>=5.17.0
pandas>=2.1.0
scikit-learn>=1.3.0
numpy>=1.24.0
openai>=1.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0